requirements.txt
absl-py==2.1.0
accelerate==1.0.1
ai4ts==0.0.3
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
axial_positional_embedding==0.2.1
backcall==0.2.0
beautifulsoup4==4.12.3
benchpots==0.3
bleach==6.1.0
bokeh==3.1.1
Bottleneck==1.4.0
cachetools==5.5.0
catboost==1.2.7
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
clarabel==0.9.0
cmake==3.30.1
CoLT5-attention==0.11.0
comm==0.2.2
contourpy==1.1.1
cryptography==43.0.0
cvxpy==1.5.2
cycler==0.12.1
Cython==3.0.11
datasets==2.21.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
dnspython==2.6.1
docopt==0.6.2
ecos==2.0.14
einops==0.8.0
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gluonts==0.15.1
google-auth==2.37.0
google-auth-oauthlib==1.0.0
graphviz==0.20.3
grpcio==1.69.0
h5py==3.11.0
huggingface-hub==0.24.3
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lightgbm==4.6.0
line_profiler==4.1.3
lit==18.1.8
local-attention==1.9.14
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
narwhals==1.27.1
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openpyxl==3.1.5
openreview-py==1.45.0
osqp==0.6.7.post3
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
patool==1.12
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.0.0
product_key_memory==0.2.11
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycrawlers==1.1.6
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pygrinder==0.6.4
PyJWT==2.9.0
pylatexenc==2.10
pymongo==4.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pypots==0.8.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qdldl==0.1.7.post5
referencing==0.35.1
reformer-pytorch==1.4.4
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
RobustPeriod==0.2.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.3
scikit-base==0.7.8
scikit-learn==1.3.2
scipy==1.10.1
scs==3.2.7.post2
seaborn==0.13.2
six==1.16.0
sktime==0.29.1
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==1.0.8
tinycss2==1.3.0
tld==0.13
tokenizers==0.19.1
toolz==0.12.1
torch==2.4.0
torchstat==0.0.7
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.1
triton==3.0.0
tsdb==0.6.2
tsfm==1.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
wrapt==1.17.0
xgboost==2.1.4
xxhash==3.4.1
xyzservices==2024.9.0
yarg==0.1.9
yarl==1.9.4
zipp==3.20.2
zmq==0.0.0
